+++ jupyterlite/check_conformance_fixtures/q_float.py
@@ -1,0 +1,21 @@
+OK_FORMAT = True
+
+test = {
+    "name": "q_float",
+    "points": None,
+    "suites": [
+        {
+            "cases": [
+                {
+                    "code": (">>> average(1, 2)\n" "1.5\n"),
+                    "hidden": False,
+                    "locked": False,
+                }
+            ],
+            "scored": True,
+            "setup": "",
+            "teardown": "",
+            "type": "doctest",
+        }
+    ],
+}