+++ jupyterlite/check_conformance_fixtures/q_multiline.py
@@ -1,0 +1,21 @@
+OK_FORMAT = True
+
+test = {
+    "name": "q_multiline",
+    "points": None,
+    "suites": [
+        {
+            "cases": [
+                {
+                    "code": (">>> print(greeting)\n" "Hello,\n" "world!\n"),
+                    "hidden": False,
+                    "locked": False,
+                }
+            ],
+            "scored": True,
+            "setup": "",
+            "teardown": "",
+            "type": "doctest",
+        }
+    ],
+}